% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_delete_replication_subnet_group}
\alias{databasemigrationservice_delete_replication_subnet_group}
\title{Deletes a subnet group}
\usage{
databasemigrationservice_delete_replication_subnet_group(
  ReplicationSubnetGroupIdentifier)
}
\arguments{
\item{ReplicationSubnetGroupIdentifier}{[required] The subnet group name of the replication instance.}
}
\description{
Deletes a subnet group.
}
\section{Request syntax}{
\preformatted{databasemigrationservice$delete_replication_subnet_group(
  ReplicationSubnetGroupIdentifier = "string"
)
}
}

\examples{
# Deletes a replication subnet group.
\donttest{databasemigrationservice$delete_replication_subnet_group(
  ReplicationSubnetGroupIdentifier = "us-west-2ab-vpc-215ds366"
)}

}
\keyword{internal}
