% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_describe_replication_tasks}
\alias{databasemigrationservice_describe_replication_tasks}
\title{Returns information about replication tasks for your account in the
current region}
\usage{
databasemigrationservice_describe_replication_tasks(Filters, MaxRecords,
  Marker, WithoutSettings)
}
\arguments{
\item{Filters}{Filters applied to the describe action.

Valid filter names: replication-task-arn \| replication-task-id \|
migration-type \| endpoint-arn \| replication-instance-arn}

\item{MaxRecords}{The maximum number of records to include in the response. If more
records exist than the specified \code{MaxRecords} value, a pagination token
called a marker is included in the response so that the remaining
results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this
parameter is specified, the response includes only records beyond the
marker, up to the value specified by \code{MaxRecords}.}

\item{WithoutSettings}{Set this flag to avoid returning setting information. Use this to reduce
overhead when settings are too large. Choose TRUE to use this flag,
otherwise choose FALSE (default).}
}
\description{
Returns information about replication tasks for your account in the
current region.
}
\section{Request syntax}{
\preformatted{databasemigrationservice$describe_replication_tasks(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string",
  WithoutSettings = TRUE|FALSE
)
}
}

\examples{
# Returns information about replication tasks for your account in the
# current region.
\donttest{databasemigrationservice$describe_replication_tasks(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string",
        "string"
      )
    )
  ),
  Marker = "",
  MaxRecords = 123L
)}

}
\keyword{internal}
