% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databasemigrationservice_operations.R
\name{databasemigrationservice_start_replication_task_assessment}
\alias{databasemigrationservice_start_replication_task_assessment}
\title{Starts the replication task assessment for unsupported data types in the
source database}
\usage{
databasemigrationservice_start_replication_task_assessment(
  ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the replication task.}
}
\description{
Starts the replication task assessment for unsupported data types in the
source database.
}
\section{Request syntax}{
\preformatted{databasemigrationservice$start_replication_task_assessment(
  ReplicationTaskArn = "string"
)
}
}

\keyword{internal}
